frontend/requirements.txt
openai>=1.0.0
gradio
requests
python-dotenv
